requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# ── HTTP client (used by openrouter, huggingface, fireworks, ollama) ───────────
httpx>=0.27.0

# ── Environment variable loading ───────────────────────────────────────────────
python-dotenv>=1.0.0

# ── LLM provider SDKs ─────────────────────────────────────────────────────────
google-generativeai>=0.7.0        # Gemini
groq>=0.9.0                       # Groq
together>=1.2.0                   # Together AI
cohere>=5.5.0                     # Cohere
replicate>=0.30.0                 # Replicate

# HuggingFace and Fireworks use plain httpx (no extra SDK needed)
# Ollama uses plain httpx (no extra SDK needed)